agent: sdeventplus: fix signal handler
`sdevent/event.hpp` was written as a copy of usable code from the early
and poor version of `sdeventplus` library.

`sdeventplus` now is stable and already present in the BMC firmware.
It also has a lot of features that are usable for this project.

This commit replaces our poor version with external full version and
fixes the crash on application termination by UNIX signal.

Signed-off-by: Alexander Filippov <a.filippov@yadro.com>